split fbs file
