+++ src/lib/schema/dalmatian_generated.ts
@@ -1,50 +1,9 @@
-// automatically generated by the FlatBuffers compiler, do not modify
-
 // tslint:disable: member-access member-ordering max-classes-per-file prefer-const no-class variable-name no-expression-statement no-object-mutation no-this readonly-keyword no-var-keyword typedef
 
 import {flatbuffers} from 'flatbuffers'
+// automatically generated by the FlatBuffers compiler, do not modify
 
-
-/**
- * @enum
- */
-export enum ArcFlag{
-  LargeArcA= 0,
-  SmallArcA= 1,
-  LargeArcB= 2,
-  SmallArcB= 3
-};
-
-/**
- * @enum
- */
-export enum Parameter{
-  NONE= 0,
-  IRIParam= 1,
-  Uint8Param= 2,
-  Uint16Param= 3,
-  Uint32Param= 4,
-  FloatParam= 5,
-  StringParam= 6,
-  DimensionParam= 7
-};
-
-/**
- * @enum
- */
-export enum Draw{
-  NONE= 0,
-  MoveTo= 1,
-  LineTo= 2,
-  Horizontal= 3,
-  Vertical= 4,
-  CubicCurve= 5,
-  SmoothCubicCurve= 6,
-  QuadraticCurve= 7,
-  SmoothQuadraticCurve= 8,
-  EllipticalArc= 9
-};
-
+import * as NS15196443072133145081 from "./ntriples_generated";
 /**
  * @enum
  */
@@ -59,557 +18,6 @@ export enum LayerData{
   Raw2DLayer32= 7
 };
 
-/**
- * @constructor
- */
-export class Dimension2D {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Dimension2D
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Dimension2D {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-width():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-height():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number width
- * @param number height
- * @returns flatbuffers.Offset
- */
-static createDimension2D(builder:flatbuffers.Builder, width: number, height: number):flatbuffers.Offset {
-  builder.prep(4, 8);
-  builder.writeInt32(height);
-  builder.writeInt32(width);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class MoveTo {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns MoveTo
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):MoveTo {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createMoveTo(builder:flatbuffers.Builder, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 8);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class LineTo {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns LineTo
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):LineTo {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createLineTo(builder:flatbuffers.Builder, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 8);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class Horizontal {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Horizontal
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Horizontal {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x
- * @returns flatbuffers.Offset
- */
-static createHorizontal(builder:flatbuffers.Builder, x: number):flatbuffers.Offset {
-  builder.prep(4, 4);
-  builder.writeInt32(x);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class Vertical {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Vertical
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Vertical {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createVertical(builder:flatbuffers.Builder, y: number):flatbuffers.Offset {
-  builder.prep(4, 4);
-  builder.writeInt32(y);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class CubicCurve {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns CubicCurve
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):CubicCurve {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x1():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-y1():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @returns number
- */
-x2():number {
-  return this.bb!.readUint32(this.bb_pos + 8);
-};
-
-/**
- * @returns number
- */
-y2():number {
-  return this.bb!.readUint32(this.bb_pos + 12);
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos + 16);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 20);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x1
- * @param number y1
- * @param number x2
- * @param number y2
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createCubicCurve(builder:flatbuffers.Builder, x1: number, y1: number, x2: number, y2: number, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 24);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  builder.writeInt32(y2);
-  builder.writeInt32(x2);
-  builder.writeInt32(y1);
-  builder.writeInt32(x1);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class SmoothCubicCurve {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns SmoothCubicCurve
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):SmoothCubicCurve {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x2():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-y2():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos + 8);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 12);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x2
- * @param number y2
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createSmoothCubicCurve(builder:flatbuffers.Builder, x2: number, y2: number, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 16);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  builder.writeInt32(y2);
-  builder.writeInt32(x2);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class QuadraticCurve {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns QuadraticCurve
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):QuadraticCurve {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x1():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-y1():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos + 8);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 12);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x1
- * @param number y1
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createQuadraticCurve(builder:flatbuffers.Builder, x1: number, y1: number, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 16);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  builder.writeInt32(y1);
-  builder.writeInt32(x1);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class SmoothQuadraticCurve {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns SmoothQuadraticCurve
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):SmoothQuadraticCurve {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createSmoothQuadraticCurve(builder:flatbuffers.Builder, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 8);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  return builder.offset();
-};
-
-}
-/**
- * @constructor
- */
-export class EllipticalArc {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns EllipticalArc
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):EllipticalArc {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @returns number
- */
-rx():number {
-  return this.bb!.readUint32(this.bb_pos);
-};
-
-/**
- * @returns number
- */
-ry():number {
-  return this.bb!.readUint32(this.bb_pos + 4);
-};
-
-/**
- * @returns number
- */
-rotate():number {
-  return this.bb!.readFloat32(this.bb_pos + 8);
-};
-
-/**
- * @returns ArcFlag
- */
-arcFlag():ArcFlag {
-  return /**  */ (this.bb!.readInt8(this.bb_pos + 12));
-};
-
-/**
- * @returns number
- */
-x():number {
-  return this.bb!.readUint32(this.bb_pos + 16);
-};
-
-/**
- * @returns number
- */
-y():number {
-  return this.bb!.readUint32(this.bb_pos + 20);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number rx
- * @param number ry
- * @param number rotate
- * @param ArcFlag arcFlag
- * @param number x
- * @param number y
- * @returns flatbuffers.Offset
- */
-static createEllipticalArc(builder:flatbuffers.Builder, rx: number, ry: number, rotate: number, arcFlag: ArcFlag, x: number, y: number):flatbuffers.Offset {
-  builder.prep(4, 24);
-  builder.writeInt32(y);
-  builder.writeInt32(x);
-  builder.pad(3);
-  builder.writeInt8(arcFlag);
-  builder.writeFloat32(rotate);
-  builder.writeInt32(ry);
-  builder.writeInt32(rx);
-  return builder.offset();
-};
-
-}
 /**
  * @constructor
  */
@@ -907,982 +315,6 @@ static endArrayOfU8(builder:flatbuffers.Builder):flatbuffers.Offset {
 /**
  * @constructor
  */
-export class IRIMeta {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns IRIMeta
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):IRIMeta {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param IRIMeta= obj
- * @returns IRIMeta
- */
-static getRootAsIRIMeta(bb:flatbuffers.ByteBuffer, obj?:IRIMeta):IRIMeta {
-  return (obj || new IRIMeta).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-prefix():string|null
-prefix(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-prefix(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-suffix():string|null
-suffix(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-suffix(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-mediaType():string|null
-mediaType(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-mediaType(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 8);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-language():string|null
-language(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-language(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 10);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startIRIMeta(builder:flatbuffers.Builder) {
-  builder.startObject(4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset prefixOffset
- */
-static addPrefix(builder:flatbuffers.Builder, prefixOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, prefixOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset suffixOffset
- */
-static addSuffix(builder:flatbuffers.Builder, suffixOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, suffixOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset mediaTypeOffset
- */
-static addMediaType(builder:flatbuffers.Builder, mediaTypeOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(2, mediaTypeOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset languageOffset
- */
-static addLanguage(builder:flatbuffers.Builder, languageOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(3, languageOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endIRIMeta(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class IRI {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns IRI
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):IRI {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param IRI= obj
- * @returns IRI
- */
-static getRootAsIRI(bb:flatbuffers.ByteBuffer, obj?:IRI):IRI {
-  return (obj || new IRI).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-path():string|null
-path(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-path(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param IRIMeta= obj
- * @returns IRIMeta|null
- */
-meta(obj?:IRIMeta):IRIMeta|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new IRIMeta).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startIRI(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset pathOffset
- */
-static addPath(builder:flatbuffers.Builder, pathOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, pathOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset metaOffset
- */
-static addMeta(builder:flatbuffers.Builder, metaOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, metaOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endIRI(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class Hyperlink {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Hyperlink
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Hyperlink {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param Hyperlink= obj
- * @returns Hyperlink
- */
-static getRootAsHyperlink(bb:flatbuffers.ByteBuffer, obj?:Hyperlink):Hyperlink {
-  return (obj || new Hyperlink).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-text():string|null
-text(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-text(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-iri(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startHyperlink(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset textOffset
- */
-static addText(builder:flatbuffers.Builder, textOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, textOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset iriOffset
- */
-static addIri(builder:flatbuffers.Builder, iriOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, iriOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endHyperlink(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class IRIParam {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns IRIParam
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):IRIParam {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param IRIParam= obj
- * @returns IRIParam
- */
-static getRootAsIRIParam(bb:flatbuffers.ByteBuffer, obj?:IRIParam):IRIParam {
-  return (obj || new IRIParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-resource(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startIRIParam(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset resourceOffset
- */
-static addResource(builder:flatbuffers.Builder, resourceOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, resourceOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endIRIParam(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class Uint8Param {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Uint8Param
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Uint8Param {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param Uint8Param= obj
- * @returns Uint8Param
- */
-static getRootAsUint8Param(bb:flatbuffers.ByteBuffer, obj?:Uint8Param):Uint8Param {
-  return (obj || new Uint8Param).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param number index
- * @returns number
- */
-value(index: number):number|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.readUint8(this.bb!.__vector(this.bb_pos + offset) + index) : 0;
-};
-
-/**
- * @returns number
- */
-valueLength():number {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
-};
-
-/**
- * @returns Uint8Array
- */
-valueArray():Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? new Uint8Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startUint8Param(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset valueOffset
- */
-static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, valueOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param Array.<number> data
- * @returns flatbuffers.Offset
- */
-static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
-  builder.startVector(1, data.length, 1);
-  for (var i = data.length - 1; i >= 0; i--) {
-    builder.addInt8(data[i]);
-  }
-  return builder.endVector();
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number numElems
- */
-static startValueVector(builder:flatbuffers.Builder, numElems:number) {
-  builder.startVector(1, numElems, 1);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endUint8Param(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class Uint16Param {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Uint16Param
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Uint16Param {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param Uint16Param= obj
- * @returns Uint16Param
- */
-static getRootAsUint16Param(bb:flatbuffers.ByteBuffer, obj?:Uint16Param):Uint16Param {
-  return (obj || new Uint16Param).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param number index
- * @returns number
- */
-value(index: number):number|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.readUint16(this.bb!.__vector(this.bb_pos + offset) + index * 2) : 0;
-};
-
-/**
- * @returns number
- */
-valueLength():number {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
-};
-
-/**
- * @returns Uint16Array
- */
-valueArray():Uint16Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? new Uint16Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startUint16Param(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset valueOffset
- */
-static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, valueOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param Array.<number> data
- * @returns flatbuffers.Offset
- */
-static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
-  builder.startVector(2, data.length, 2);
-  for (var i = data.length - 1; i >= 0; i--) {
-    builder.addInt16(data[i]);
-  }
-  return builder.endVector();
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number numElems
- */
-static startValueVector(builder:flatbuffers.Builder, numElems:number) {
-  builder.startVector(2, numElems, 2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endUint16Param(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class Uint32Param {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns Uint32Param
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):Uint32Param {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param Uint32Param= obj
- * @returns Uint32Param
- */
-static getRootAsUint32Param(bb:flatbuffers.ByteBuffer, obj?:Uint32Param):Uint32Param {
-  return (obj || new Uint32Param).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param number index
- * @returns number
- */
-value(index: number):number|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.readUint32(this.bb!.__vector(this.bb_pos + offset) + index * 4) : 0;
-};
-
-/**
- * @returns number
- */
-valueLength():number {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
-};
-
-/**
- * @returns Uint32Array
- */
-valueArray():Uint32Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? new Uint32Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startUint32Param(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset valueOffset
- */
-static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, valueOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param Array.<number> data
- * @returns flatbuffers.Offset
- */
-static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
-  builder.startVector(4, data.length, 4);
-  for (var i = data.length - 1; i >= 0; i--) {
-    builder.addInt32(data[i]);
-  }
-  return builder.endVector();
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number numElems
- */
-static startValueVector(builder:flatbuffers.Builder, numElems:number) {
-  builder.startVector(4, numElems, 4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endUint32Param(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class FloatParam {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns FloatParam
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):FloatParam {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param FloatParam= obj
- * @returns FloatParam
- */
-static getRootAsFloatParam(bb:flatbuffers.ByteBuffer, obj?:FloatParam):FloatParam {
-  return (obj || new FloatParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param number index
- * @returns number
- */
-value(index: number):number|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.readFloat32(this.bb!.__vector(this.bb_pos + offset) + index * 4) : 0;
-};
-
-/**
- * @returns number
- */
-valueLength():number {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
-};
-
-/**
- * @returns Float32Array
- */
-valueArray():Float32Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? new Float32Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startFloatParam(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset valueOffset
- */
-static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, valueOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param Array.<number> data
- * @returns flatbuffers.Offset
- */
-static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
-  builder.startVector(4, data.length, 4);
-  for (var i = data.length - 1; i >= 0; i--) {
-    builder.addFloat32(data[i]);
-  }
-  return builder.endVector();
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param number numElems
- */
-static startValueVector(builder:flatbuffers.Builder, numElems:number) {
-  builder.startVector(4, numElems, 4);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endFloatParam(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class StringParam {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns StringParam
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):StringParam {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param StringParam= obj
- * @returns StringParam
- */
-static getRootAsStringParam(bb:flatbuffers.ByteBuffer, obj?:StringParam):StringParam {
-  return (obj || new StringParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param flatbuffers.Encoding= optionalEncoding
- * @returns string|Uint8Array|null
- */
-value():string|null
-value(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
-value(optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startStringParam(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset valueOffset
- */
-static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, valueOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endStringParam(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
-export class DimensionParam {
-  bb: flatbuffers.ByteBuffer|null = null;
-
-  bb_pos:number = 0;
-/**
- * @param number i
- * @param flatbuffers.ByteBuffer bb
- * @returns DimensionParam
- */
-__init(i:number, bb:flatbuffers.ByteBuffer):DimensionParam {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param flatbuffers.ByteBuffer bb
- * @param DimensionParam= obj
- * @returns DimensionParam
- */
-static getRootAsDimensionParam(bb:flatbuffers.ByteBuffer, obj?:DimensionParam):DimensionParam {
-  return (obj || new DimensionParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param IRI= obj
- * @returns IRI|null
- */
-predicate(obj?:IRI):IRI|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param Dimension2D= obj
- * @returns Dimension2D|null
- */
-value(obj?:Dimension2D):Dimension2D|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new Dimension2D).__init(this.bb_pos + offset, this.bb!) : null;
-};
-
-/**
- * @param flatbuffers.Builder builder
- */
-static startDimensionParam(builder:flatbuffers.Builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset predicateOffset
- */
-static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, predicateOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset valueOffset
- */
-static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
-  builder.addFieldStruct(1, valueOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @returns flatbuffers.Offset
- */
-static endDimensionParam(builder:flatbuffers.Builder):flatbuffers.Offset {
-  var offset = builder.endObject();
-  return offset;
-};
-
-}
-/**
- * @constructor
- */
 export class Metadata {
   bb: flatbuffers.ByteBuffer|null = null;
 
@@ -1911,18 +343,18 @@ static getRootAsMetadata(bb:flatbuffers.ByteBuffer, obj?:Metadata):Metadata {
  * @param IRI= obj
  * @returns IRI|null
  */
-identifier(obj?:IRI):IRI|null {
+identifier(obj?:NS15196443072133145081.IRI):NS15196443072133145081.IRI|null {
   var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param IRI= obj
  * @returns IRI|null
  */
-version(obj?:IRI):IRI|null {
+version(obj?:NS15196443072133145081.IRI):NS15196443072133145081.IRI|null {
   var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -1940,18 +372,18 @@ language(optionalEncoding?:any):string|Uint8Array|null {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-title(obj?:Hyperlink):Hyperlink|null {
+title(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 10);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-description(obj?:Hyperlink):Hyperlink|null {
+description(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 12);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -1959,9 +391,9 @@ description(obj?:Hyperlink):Hyperlink|null {
  * @param Hyperlink= obj
  * @returns Hyperlink
  */
-creator(index: number, obj?:Hyperlink):Hyperlink|null {
+creator(index: number, obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 14);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
@@ -1977,9 +409,9 @@ creatorLength():number {
  * @param Hyperlink= obj
  * @returns Hyperlink
  */
-contributor(index: number, obj?:Hyperlink):Hyperlink|null {
+contributor(index: number, obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 16);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
@@ -1995,9 +427,9 @@ contributorLength():number {
  * @param Hyperlink= obj
  * @returns Hyperlink
  */
-publisher(index: number, obj?:Hyperlink):Hyperlink|null {
+publisher(index: number, obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 18);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
@@ -2012,27 +444,27 @@ publisherLength():number {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-license(obj?:Hyperlink):Hyperlink|null {
+license(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 20);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-rights(obj?:Hyperlink):Hyperlink|null {
+rights(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 22);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-attribution(obj?:Hyperlink):Hyperlink|null {
+attribution(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 24);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -2062,9 +494,9 @@ modified(optionalEncoding?:any):string|Uint8Array|null {
  * @param IRIParam= obj
  * @returns IRIParam
  */
-related(index: number, obj?:IRIParam):IRIParam|null {
+related(index: number, obj?:NS15196443072133145081.IRIParam):NS15196443072133145081.IRIParam|null {
   var offset = this.bb!.__offset(this.bb_pos, 30);
-  return offset ? (obj || new IRIParam).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRIParam).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
@@ -2319,18 +751,18 @@ static getRootAsColor(bb:flatbuffers.ByteBuffer, obj?:Color):Color {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-name(obj?:Hyperlink):Hyperlink|null {
+name(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-colorSpace(obj?:Hyperlink):Hyperlink|null {
+colorSpace(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -2416,18 +848,18 @@ static getRootAsProcessor(bb:flatbuffers.ByteBuffer, obj?:Processor):Processor {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-name(obj?:Hyperlink):Hyperlink|null {
+name(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-version(obj?:Hyperlink):Hyperlink|null {
+version(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -2675,9 +1107,9 @@ static getRootAsEffect(bb:flatbuffers.ByteBuffer, obj?:Effect):Effect {
  * @param IRI= obj
  * @returns IRI|null
  */
-event(obj?:IRI):IRI|null {
+event(obj?:NS15196443072133145081.IRI):NS15196443072133145081.IRI|null {
   var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -2751,7 +1183,7 @@ static addEvent(builder:flatbuffers.Builder, eventOffset:flatbuffers.Offset) {
  * @param flatbuffers.Builder builder
  * @param flatbuffers.Offset parameterTypeOffset
  */
-static addParameterType(builder:flatbuffers.Builder, parameterTypeOffset:flatbuffers.Offset) {
+static addParameterType(builder:flatbuffers.Builder, parameterTypeOffset:NS15196443072133145081.flatbuffers.Offset) {
   builder.addFieldOffset(1, parameterTypeOffset, 0);
 };
 
@@ -2780,7 +1212,7 @@ static startParameterTypeVector(builder:flatbuffers.Builder, numElems:number) {
  * @param flatbuffers.Builder builder
  * @param flatbuffers.Offset parameterOffset
  */
-static addParameter(builder:flatbuffers.Builder, parameterOffset:flatbuffers.Offset) {
+static addParameter(builder:flatbuffers.Builder, parameterOffset:NS15196443072133145081.flatbuffers.Offset) {
   builder.addFieldOffset(2, parameterOffset, 0);
 };
 
@@ -2854,27 +1286,27 @@ static getRootAsShape(bb:flatbuffers.ByteBuffer, obj?:Shape):Shape {
  * @param IRI= obj
  * @returns IRI|null
  */
-identifier(obj?:IRI):IRI|null {
+identifier(obj?:NS15196443072133145081.IRI):NS15196443072133145081.IRI|null {
   var offset = this.bb!.__offset(this.bb_pos, 4);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-title(obj?:Hyperlink):Hyperlink|null {
+title(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-description(obj?:Hyperlink):Hyperlink|null {
+description(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 8);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -2882,9 +1314,9 @@ description(obj?:Hyperlink):Hyperlink|null {
  * @param IRIParam= obj
  * @returns IRIParam
  */
-related(index: number, obj?:IRIParam):IRIParam|null {
+related(index: number, obj?:NS15196443072133145081.IRIParam):NS15196443072133145081.IRIParam|null {
   var offset = this.bb!.__offset(this.bb_pos, 10);
-  return offset ? (obj || new IRIParam).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRIParam).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
@@ -3020,7 +1452,7 @@ static startRelatedVector(builder:flatbuffers.Builder, numElems:number) {
  * @param flatbuffers.Builder builder
  * @param flatbuffers.Offset drawingTypeOffset
  */
-static addDrawingType(builder:flatbuffers.Builder, drawingTypeOffset:flatbuffers.Offset) {
+static addDrawingType(builder:flatbuffers.Builder, drawingTypeOffset:NS14927261283899465511.flatbuffers.Offset) {
   builder.addFieldOffset(4, drawingTypeOffset, 0);
 };
 
@@ -3049,7 +1481,7 @@ static startDrawingTypeVector(builder:flatbuffers.Builder, numElems:number) {
  * @param flatbuffers.Builder builder
  * @param flatbuffers.Offset drawingOffset
  */
-static addDrawing(builder:flatbuffers.Builder, drawingOffset:flatbuffers.Offset) {
+static addDrawing(builder:flatbuffers.Builder, drawingOffset:NS14927261283899465511.flatbuffers.Offset) {
   builder.addFieldOffset(5, drawingOffset, 0);
 };
 
@@ -3187,9 +1619,9 @@ parameterLength():number {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-encoding(obj?:Hyperlink):Hyperlink|null {
+encoding(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 8);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -3203,7 +1635,7 @@ static startRawLayerMeta(builder:flatbuffers.Builder) {
  * @param flatbuffers.Builder builder
  * @param flatbuffers.Offset parameterTypeOffset
  */
-static addParameterType(builder:flatbuffers.Builder, parameterTypeOffset:flatbuffers.Offset) {
+static addParameterType(builder:flatbuffers.Builder, parameterTypeOffset:NS15196443072133145081.flatbuffers.Offset) {
   builder.addFieldOffset(0, parameterTypeOffset, 0);
 };
 
@@ -3232,7 +1664,7 @@ static startParameterTypeVector(builder:flatbuffers.Builder, numElems:number) {
  * @param flatbuffers.Builder builder
  * @param flatbuffers.Offset parameterOffset
  */
-static addParameter(builder:flatbuffers.Builder, parameterOffset:flatbuffers.Offset) {
+static addParameter(builder:flatbuffers.Builder, parameterOffset:NS15196443072133145081.flatbuffers.Offset) {
   builder.addFieldOffset(1, parameterOffset, 0);
 };
 
@@ -4083,18 +2515,18 @@ data<T extends flatbuffers.Table>(obj:T):T|null {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-format(obj?:Hyperlink):Hyperlink|null {
+format(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 8);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-description(obj?:Hyperlink):Hyperlink|null {
+description(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 10);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -4242,18 +2674,18 @@ layerLength():number {
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-medium(obj?:Hyperlink):Hyperlink|null {
+medium(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
  * @param Hyperlink= obj
  * @returns Hyperlink|null
  */
-format(obj?:Hyperlink):Hyperlink|null {
+format(obj?:NS15196443072133145081.Hyperlink):NS15196443072133145081.Hyperlink|null {
   var offset = this.bb!.__offset(this.bb_pos, 8);
-  return offset ? (obj || new Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Hyperlink).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
 };
 
 /**
@@ -4261,9 +2693,9 @@ format(obj?:Hyperlink):Hyperlink|null {
  * @param IRI= obj
  * @returns IRI
  */
-style(index: number, obj?:IRI):IRI|null {
+style(index: number, obj?:NS15196443072133145081.IRI):NS15196443072133145081.IRI|null {
   var offset = this.bb!.__offset(this.bb_pos, 10);
-  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRI).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
@@ -4296,9 +2728,9 @@ colorLength():number {
  * @param Dimension2D= obj
  * @returns Dimension2D|null
  */
-dimension(obj?:Dimension2D):Dimension2D|null {
+dimension(obj?:NS15196443072133145081.Dimension2D):NS15196443072133145081.Dimension2D|null {
   var offset = this.bb!.__offset(this.bb_pos, 14);
-  return offset ? (obj || new Dimension2D).__init(this.bb_pos + offset, this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.Dimension2D).__init(this.bb_pos + offset, this.bb!) : null;
 };
 
 /**
@@ -4314,9 +2746,9 @@ dpi():number {
  * @param IRIParam= obj
  * @returns IRIParam
  */
-related(index: number, obj?:IRIParam):IRIParam|null {
+related(index: number, obj?:NS15196443072133145081.IRIParam):NS15196443072133145081.IRIParam|null {
   var offset = this.bb!.__offset(this.bb_pos, 18);
-  return offset ? (obj || new IRIParam).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
+  return offset ? (obj || new NS15196443072133145081.IRIParam).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
 /**
