+++ src/lib/schema/ntriples_generated.ts
@@ -1,0 +1,1041 @@
+// tslint:disable: member-access member-ordering max-classes-per-file prefer-const no-class variable-name no-expression-statement no-object-mutation no-this readonly-keyword no-var-keyword typedef
+
+import {flatbuffers} from 'flatbuffers'
+// automatically generated by the FlatBuffers compiler, do not modify
+
+/**
+ * @enum
+ */
+export enum Parameter{
+  NONE= 0,
+  IRIParam= 1,
+  Uint8Param= 2,
+  Uint16Param= 3,
+  Uint32Param= 4,
+  FloatParam= 5,
+  StringParam= 6,
+  DimensionParam= 7
+};
+
+/**
+ * @constructor
+ */
+export class Dimension2D {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns Dimension2D
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):Dimension2D {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @returns number
+ */
+width():number {
+  return this.bb!.readUint32(this.bb_pos);
+};
+
+/**
+ * @returns number
+ */
+height():number {
+  return this.bb!.readUint32(this.bb_pos + 4);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number width
+ * @param number height
+ * @returns flatbuffers.Offset
+ */
+static createDimension2D(builder:flatbuffers.Builder, width: number, height: number):flatbuffers.Offset {
+  builder.prep(4, 8);
+  builder.writeInt32(height);
+  builder.writeInt32(width);
+  return builder.offset();
+};
+
+}
+/**
+ * @constructor
+ */
+export class IRIMeta {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns IRIMeta
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):IRIMeta {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param IRIMeta= obj
+ * @returns IRIMeta
+ */
+static getRootAsIRIMeta(bb:flatbuffers.ByteBuffer, obj?:IRIMeta):IRIMeta {
+  return (obj || new IRIMeta).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+prefix():string|null
+prefix(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+prefix(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+suffix():string|null
+suffix(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+suffix(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+mediaType():string|null
+mediaType(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+mediaType(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+language():string|null
+language(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+language(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startIRIMeta(builder:flatbuffers.Builder) {
+  builder.startObject(4);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset prefixOffset
+ */
+static addPrefix(builder:flatbuffers.Builder, prefixOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, prefixOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset suffixOffset
+ */
+static addSuffix(builder:flatbuffers.Builder, suffixOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, suffixOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset mediaTypeOffset
+ */
+static addMediaType(builder:flatbuffers.Builder, mediaTypeOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(2, mediaTypeOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset languageOffset
+ */
+static addLanguage(builder:flatbuffers.Builder, languageOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(3, languageOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endIRIMeta(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class IRI {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns IRI
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):IRI {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param IRI= obj
+ * @returns IRI
+ */
+static getRootAsIRI(bb:flatbuffers.ByteBuffer, obj?:IRI):IRI {
+  return (obj || new IRI).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+path():string|null
+path(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+path(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param IRIMeta= obj
+ * @returns IRIMeta|null
+ */
+meta(obj?:IRIMeta):IRIMeta|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? (obj || new IRIMeta).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startIRI(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset pathOffset
+ */
+static addPath(builder:flatbuffers.Builder, pathOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, pathOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset metaOffset
+ */
+static addMeta(builder:flatbuffers.Builder, metaOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, metaOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endIRI(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class Hyperlink {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns Hyperlink
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):Hyperlink {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param Hyperlink= obj
+ * @returns Hyperlink
+ */
+static getRootAsHyperlink(bb:flatbuffers.ByteBuffer, obj?:Hyperlink):Hyperlink {
+  return (obj || new Hyperlink).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+text():string|null
+text(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+text(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+iri(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startHyperlink(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset textOffset
+ */
+static addText(builder:flatbuffers.Builder, textOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, textOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset iriOffset
+ */
+static addIri(builder:flatbuffers.Builder, iriOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, iriOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endHyperlink(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class IRIParam {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns IRIParam
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):IRIParam {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param IRIParam= obj
+ * @returns IRIParam
+ */
+static getRootAsIRIParam(bb:flatbuffers.ByteBuffer, obj?:IRIParam):IRIParam {
+  return (obj || new IRIParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+resource(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startIRIParam(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset resourceOffset
+ */
+static addResource(builder:flatbuffers.Builder, resourceOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, resourceOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endIRIParam(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class Uint8Param {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns Uint8Param
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):Uint8Param {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param Uint8Param= obj
+ * @returns Uint8Param
+ */
+static getRootAsUint8Param(bb:flatbuffers.ByteBuffer, obj?:Uint8Param):Uint8Param {
+  return (obj || new Uint8Param).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+value(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.readUint8(this.bb!.__vector(this.bb_pos + offset) + index) : 0;
+};
+
+/**
+ * @returns number
+ */
+valueLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Uint8Array
+ */
+valueArray():Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? new Uint8Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startUint8Param(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset valueOffset
+ */
+static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, valueOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(1, data.length, 1);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addInt8(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startValueVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(1, numElems, 1);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endUint8Param(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class Uint16Param {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns Uint16Param
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):Uint16Param {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param Uint16Param= obj
+ * @returns Uint16Param
+ */
+static getRootAsUint16Param(bb:flatbuffers.ByteBuffer, obj?:Uint16Param):Uint16Param {
+  return (obj || new Uint16Param).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+value(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.readUint16(this.bb!.__vector(this.bb_pos + offset) + index * 2) : 0;
+};
+
+/**
+ * @returns number
+ */
+valueLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Uint16Array
+ */
+valueArray():Uint16Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? new Uint16Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startUint16Param(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset valueOffset
+ */
+static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, valueOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(2, data.length, 2);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addInt16(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startValueVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(2, numElems, 2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endUint16Param(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class Uint32Param {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns Uint32Param
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):Uint32Param {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param Uint32Param= obj
+ * @returns Uint32Param
+ */
+static getRootAsUint32Param(bb:flatbuffers.ByteBuffer, obj?:Uint32Param):Uint32Param {
+  return (obj || new Uint32Param).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+value(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.readUint32(this.bb!.__vector(this.bb_pos + offset) + index * 4) : 0;
+};
+
+/**
+ * @returns number
+ */
+valueLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Uint32Array
+ */
+valueArray():Uint32Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? new Uint32Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startUint32Param(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset valueOffset
+ */
+static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, valueOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(4, data.length, 4);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addInt32(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startValueVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(4, numElems, 4);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endUint32Param(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class FloatParam {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns FloatParam
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):FloatParam {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param FloatParam= obj
+ * @returns FloatParam
+ */
+static getRootAsFloatParam(bb:flatbuffers.ByteBuffer, obj?:FloatParam):FloatParam {
+  return (obj || new FloatParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+value(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.readFloat32(this.bb!.__vector(this.bb_pos + offset) + index * 4) : 0;
+};
+
+/**
+ * @returns number
+ */
+valueLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Float32Array
+ */
+valueArray():Float32Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? new Float32Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startFloatParam(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset valueOffset
+ */
+static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, valueOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createValueVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(4, data.length, 4);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addFloat32(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startValueVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(4, numElems, 4);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endFloatParam(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class StringParam {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns StringParam
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):StringParam {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param StringParam= obj
+ * @returns StringParam
+ */
+static getRootAsStringParam(bb:flatbuffers.ByteBuffer, obj?:StringParam):StringParam {
+  return (obj || new StringParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+value():string|null
+value(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+value(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startStringParam(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset valueOffset
+ */
+static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, valueOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endStringParam(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
+/**
+ * @constructor
+ */
+export class DimensionParam {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns DimensionParam
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):DimensionParam {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param DimensionParam= obj
+ * @returns DimensionParam
+ */
+static getRootAsDimensionParam(bb:flatbuffers.ByteBuffer, obj?:DimensionParam):DimensionParam {
+  return (obj || new DimensionParam).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param IRI= obj
+ * @returns IRI|null
+ */
+predicate(obj?:IRI):IRI|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new IRI).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
+ * @param Dimension2D= obj
+ * @returns Dimension2D|null
+ */
+value(obj?:Dimension2D):Dimension2D|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? (obj || new Dimension2D).__init(this.bb_pos + offset, this.bb!) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startDimensionParam(builder:flatbuffers.Builder) {
+  builder.startObject(2);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset predicateOffset
+ */
+static addPredicate(builder:flatbuffers.Builder, predicateOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, predicateOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset valueOffset
+ */
+static addValue(builder:flatbuffers.Builder, valueOffset:flatbuffers.Offset) {
+  builder.addFieldStruct(1, valueOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endDimensionParam(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+}
